requirements.txt
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
future==1.0.0
idna==3.10
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
pycryptodomex==3.21.0
pyjwkest==1.4.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-taiga==1.3.2
requests==2.32.3
six==1.17.0
urllib3==2.3.0
